+++ src/util/Requester.ts
@@ -1,0 +1,4 @@
+export interface Requester {
+  connect(): Promise<void>;
+  disconnect(): Promise<void>;
+}
